moved image parsing function to prefect.docker.py as utility
